refactor: split the util helpers to a subdirectory
