+++ src/annotate.ts
@@ -2,7 +2,7 @@ import debug from 'debug';
 import execa from 'execa';
 import _ from 'lodash';
 import Config from './config';
-import { plurals } from './util';
+import { plurals } from './util/helper';
 
 const log = debug('monofo:annotate');
