+++ src/artifacts/compression/lz4.ts
@@ -1,7 +1,7 @@
 import stream from 'stream';
 import debug from 'debug';
 import execa from 'execa';
-import { stdinWritable } from '../../util';
+import { stdinWritable } from '../../util/exec';
 import { tar } from './tar';
 
 const log = debug('monofo:artifact:lz4');
