+++ src/artifacts/inflate.ts
@@ -3,7 +3,7 @@ import stream from 'stream';
 import { promisify } from 'util';
 import debug from 'debug';
 import execa from 'execa';
-import { stdinWritable } from '../util';
+import { stdinWritable } from '../util/exec';
 import { inflateDesync, isDesyncEnabled } from './compression/desync';
 import { tar } from './compression/tar';
 import { Artifact } from './model';
