+++ src/util/async.ts
@@ -1,3 +1,18 @@
+export function mapAsync<I, O>(
+  array: I[],
+  callback: (value: I, index: number, array: I[]) => Promise<O>
+): Promise<O[]> {
+  return Promise.all(array.map(callback));
+}
+
+export async function filterAsync<T>(
+  array: T[],
+  callback: (value: T, index: number, array: T[]) => Promise<boolean>
+): Promise<T[]> {
+  const filterMap = await mapAsync(array, callback);
+  return array.filter((_, index) => filterMap[index]);
+}
+
 /**
  * Given an async iterable of string chunks (that may not end at separator boundaries), split the stream into an iterable
  * of separated parts, between the separator (and excluding it)
